+++ src/wrapt/__init__.py
@@ -1,4 +1,4 @@
-__version_info__ = ('1', '11', '2')
+__version_info__ = ('1', '12', '0')
 __version__ = '.'.join(__version_info__)
 
 from .wrappers import (ObjectProxy, CallableObjectProxy, FunctionWrapper,